FST 83 changes from review
