+++ package.json
@@ -2,6 +2,7 @@
   "name": "fint-skjema-integration-frontend",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:8080/api",
   "dependencies": {
     "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
