+++ src/features/log/Log.tsx
@@ -4,9 +4,9 @@ import {Typography} from "@mui/material";
 
 const Log: React.FunctionComponent<RouteComponentProps<any>> = () => {
     return (
-        <div>
+        <>
             <Typography>Logg</Typography>
-        </div>
+        </>
     );
 }
