+++ src/features/support/support.tsx
@@ -4,9 +4,9 @@ import {Typography} from "@mui/material";
 
 const Support: React.FunctionComponent<RouteComponentProps<any>> = () => {
     return (
-        <div>
+        <>
             <Typography>Opprett support sak</Typography>
-        </div>
+        </>
     );
 }
