--- src/http-common.ts
@@ -1,9 +1,0 @@
-import axios from "axios";
-
-export default axios.create({
-    // TODO: change baseURL
-    baseURL: "http://localhost:8080/api",
-    headers: {
-        "Content-type": "application/json"
-    }
-});
